requirements.txt
pytube==12.0.0
Pillow==9.0.1